requirements.txt
Django>=2.2.12,<3.0.0
pandas==1.1.2
django-csp>=3.5
django-cors-headers>=3.2.1
gunicorn>=20.0.4
requests-toolbelt>=0.9.1
tco>=1.2.1
py-eureka-client==0.7.2
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
contextlib2==0.5.5
djangorestframework
idna==2.9
kafka-python==2.0.1
py-zipkin==0.18.7
python-dateutil==2.8.1
python-json-logger==0.1.11
pytz==2019.3
requests==2.23.0
schema==0.7.1
six==1.14.0
sqlparse==0.3.0
thriftpy2==0.4.10
urllib3==1.25.8
xlrd==1.2.0
